pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-email-triage-env"
version = "0.1.0"
description = "Email Triage Environment for OpenEnv - AI agent email classification, prioritization, routing, and response drafting"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "BSD-3-Clause"}

dependencies = [
    # Core OpenEnv dependencies (required for server functionality)
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
]
baseline = [
    "openai>=1.30.0",
]

[project.scripts]
server = "email_triage_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["email_triage_env", "email_triage_env.server"]
package-dir = { "email_triage_env" = ".", "email_triage_env.server" = "server" }
package-data = { "email_triage_env" = ["data/*.json"] }
